Refactor and simplify cli code
In particular, now using `click` for command line options. This also
opens the door to multiple commands in the future that can be called
via `scripts/gd`, if warranted (for now I've consolidated everything
into a single `build-metadata` command.)
