+++ .circleci/config.yml
@@ -56,11 +56,8 @@ jobs:
           name: Install package locally
           command: venv/bin/pip install -e .
       - run:
-          name: Build glean metadata
-          command: scripts/build-glean-metadata
-      - run:
-          name: Build legacy firefox metadata
-          command: scripts/build-legacy-metadata
+          name: Build metadata
+          command: scripts/gd build-metadata
       - run: cp -r public /tmp/workspace
       - persist_to_workspace:
           root: /tmp/workspace
