+++ netlify.toml
@@ -2,8 +2,7 @@
 command = """
     set -e
     pip install -e .
-    ./scripts/build-glean-metadata
-    ./scripts/build-legacy-metadata
+    ./scripts/gd build-metadata
     # glean.js currently requires a virtual environment
     python -mvenv venv --without-pip
     wget https://bootstrap.pypa.io/get-pip.py
