+++ requirements.in
@@ -1,5 +1,6 @@
 autoflake==1.4
 black==21.12b0
+click==8.0.3
 furl==2.1.3
 isort==5.10.1
 Jinja2==3.0.3
