+++ requirements.txt
@@ -29,6 +29,7 @@ click==8.0.3 \
     --hash=sha256:353f466495adaeb40b6b5f592f9f91cb22372351c84caeb068132442a4518ef3 \
     --hash=sha256:410e932b050f5eed773c4cda94de75971c89cdb3155a72a0831139a79e5ecb5b
     # via
+    #   -r requirements.in
     #   black
     #   pip-tools
 flake8==4.0.1 \
